refactor(tests): consolidate tests and remove duplicates
- Merge test_sort_command.py into test_sort.py
- Remove 8 duplicate tests from test_find.py (covered by test_integration.py)
- Remove 2 redundant locale tests (already parameterized)
- Move create_requirements_file helper to conftest.py as shared fixture
- Replace local runner fixture with cli_runner from conftest.py

Test count reduced from 177 to 167 (-10 duplicates)
Total test lines reduced from 2874 to 2696 (-178 lines)
