+++ tests/test_preview_output.py
@@ -1,33 +1,11 @@
-from pathlib import Path
-
-import pytest
-from click.testing import CliRunner
-
 from src.main import cli
 
 
-@pytest.fixture
-def runner():
-    """Click test runner."""
-    return CliRunner()
-
-
-def create_requirements_file(base_path: Path, subdir: str, content: str) -> Path:
-    """Create a requirements.txt file with given content."""
-    project_dir = base_path / subdir
-    project_dir.mkdir(exist_ok=True)
-
-    file_path = project_dir / "requirements.txt"
-    file_path.write_text(content)
-
-    return file_path
-
-
-def test_update_preview_output_format(runner, tmp_path):
+def test_update_preview_output_format(cli_runner, tmp_path, create_requirements_file):
     """Test that update command uses consistent output format in preview mode."""
     create_requirements_file(tmp_path, "project", "django==3.0\nrequests==2.25.1\n")
 
-    result = runner.invoke(
+    result = cli_runner.invoke(
         cli, ["update", "requests", "2.30.0", "--preview", str(tmp_path)]
     )
 
@@ -38,11 +16,11 @@ def test_update_preview_output_format(runner, tmp_path):
     assert "requests==2.30.0" in result.stdout
 
 
-def test_add_preview_output_format(runner, tmp_path):
+def test_add_preview_output_format(cli_runner, tmp_path, create_requirements_file):
     """Test that add command uses consistent output format in preview mode."""
     create_requirements_file(tmp_path, "project", "django==3.0\n")
 
-    result = runner.invoke(cli, ["add", "pytest", "--preview", str(tmp_path)])
+    result = cli_runner.invoke(cli, ["add", "pytest", "--preview", str(tmp_path)])
 
     # Should show file path
     assert "requirements.txt" in result.stdout
@@ -51,11 +29,11 @@ def test_add_preview_output_format(runner, tmp_path):
     assert "pytest" in result.stdout
 
 
-def test_remove_preview_output_format(runner, tmp_path):
+def test_remove_preview_output_format(cli_runner, tmp_path, create_requirements_file):
     """Test that remove command uses consistent output format in preview mode."""
     create_requirements_file(tmp_path, "project", "django==3.0\nrequests==2.25.1\n")
 
-    result = runner.invoke(cli, ["remove", "requests", "--preview", str(tmp_path)])
+    result = cli_runner.invoke(cli, ["remove", "requests", "--preview", str(tmp_path)])
 
     # Should show file path
     assert "requirements.txt" in result.stdout
@@ -65,11 +43,11 @@ def test_remove_preview_output_format(runner, tmp_path):
     assert "requests==2.25.1" not in result.stdout
 
 
-def test_sort_preview_output_format(runner, tmp_path):
+def test_sort_preview_output_format(cli_runner, tmp_path, create_requirements_file):
     """Test that sort command uses consistent output format in preview mode."""
     create_requirements_file(tmp_path, "project", "requests==2.25.1\ndjango==3.0\n")
 
-    result = runner.invoke(cli, ["sort", "--preview", str(tmp_path)])
+    result = cli_runner.invoke(cli, ["sort", "--preview", str(tmp_path)])
 
     # Should show file path
     assert "requirements.txt" in result.stdout
@@ -88,7 +66,7 @@ def test_sort_preview_output_format(runner, tmp_path):
     assert "requests==2.25.1" in result.stdout
 
 
-def test_all_preview_commands_consistent_format(runner, tmp_path):
+def test_all_preview_commands_consistent_format(cli_runner, tmp_path, create_requirements_file):
     """Test that all preview commands use consistent output format."""
     create_requirements_file(tmp_path, "project", "requests==2.25.1\ndjango==3.0\n")
 
@@ -101,7 +79,7 @@ def test_all_preview_commands_consistent_format(runner, tmp_path):
     ]
 
     for cmd in commands:
-        result = runner.invoke(cli, cmd)
+        result = cli_runner.invoke(cli, cmd)
 
         # All should show file path consistently
         assert "requirements.txt" in result.stdout, (
